% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IniFunctions.R
\name{MakeControlfile}
\alias{MakeControlfile}
\title{Contruct a Controlfile for simulations}
\usage{
MakeControlfile(
  IniPath = "",
  CtrlPath = "",
  CtrlFileName = "Ctrl.fn",
  HarvestYears,
  Treatment1,
  Treatment2,
  Treatment3
)
}
\arguments{
\item{IniPath}{path to the directory where the control file should be stored}

\item{CtrlPath}{path to the directory where the control file should be stored}

\item{CtrlFileName}{name of the control file}

\item{HarvestYears}{array of integers with the years of the simulations}

\item{Treatment1}{array of strings with the names of treatment level 1}

\item{Treatment2}{array of strings with the names of treatment level 2}

\item{Treatment3}{array of strings with the names of treatment level 3}
}
\value{
nothing
}
\description{
Contruct a Controlfile for simulations
}
