% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IniFunctions.R
\name{SetTimeArrays}
\alias{SetTimeArrays}
\title{create time arrays for simulations}
\usage{
SetTimeArrays(
  FirstHarvestYear,
  LastHarvestYear,
  Wintercrop = TRUE,
  SimStart,
  SimEnd,
  Sowingdate,
  HarvestDate
)
}
\arguments{
\item{FirstHarvestYear}{First harvest year of the simulation \link{integer}}

\item{LastHarvestYear}{Last harvest year of the simulation \link{integer}}

\item{Wintercrop}{logical, is the crop a winter crop? \link{logical}}

\item{SimStart}{start date of the simulation "YYYY-MM-DD"}

\item{SimEnd}{end date of the simulation "YYYY-MM-DD"}

\item{Sowingdate}{sowing date of the crop "YYYY-MM-DD"}

\item{HarvestDate}{harvest date of the crop "YYYY-MM-DD"}
}
\value{
list with time arrays for simulations in Excel year format
}
\description{
create time arrays for simulations
}
