% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IniFunctions.R
\name{SetupDirs}
\alias{SetupDirs}
\title{Setup of directories for a simulation project}
\usage{
SetupDirs(ProjectName, TrialName = "", ModelDrive = "Q:", OutputDrive = "P:")
}
\arguments{
\item{ProjectName}{Name of the project \link{string} eg. "MaizeEvaluation"}

\item{ModelDrive}{Drive letter for the model \link{string}}

\item{OutputDrive}{Drive letter for the output \link{string}}

\item{Trial}{Name of the trial \link{string} if Trial is not empty, a subdirectory with the name of the trial will be created in the data and ini directories}
}
\value{
list with directories for the project
}
\description{
Setup of directories for a simulation project
}
\examples{
Setupirs("TestProject", "Trial1")
}
