% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IniFunctions.R
\name{WriteHumeDataFile}
\alias{WriteHumeDataFile}
\title{Title}
\usage{
WriteHumeDataFile(df, fn, namesHUME, unitsHUME)
}
\arguments{
\item{df}{data frame with the data to be written to the HUME file}

\item{fn}{file name of the HUME file to be written}

\item{namesHUME}{vector with the names of the columns in the HUME file}

\item{unitsHUME}{vector with the units of the columns in the HUME file}
}
\description{
Title
}
